+++ src/lightly_train/_task_models/dinov3_eomt_semantic_segmentation/transforms.py
@@ -24,6 +24,7 @@ from lightly_train._transforms.transform import (
     ScaleJitterArgs,
     SmallestMaxSizeArgs,
 )
+from lightly_train.types import ImageSizeTuple
 
 
 class DINOv3EoMTSemanticSegmentationColorJitterArgs(ColorJitterArgs):
@@ -73,7 +74,7 @@ class DINOv3EoMTSemanticSegmentationTrainTransformArgs(
     """
 
     # TODO(Guarin, 08/25): Check if we should change default to 512.
-    image_size: tuple[int, int] | Literal["auto"] = "auto"
+    image_size: ImageSizeTuple | Literal["auto"] = "auto"
     channel_drop: ChannelDropArgs | None = None
     num_channels: int | Literal["auto"] = "auto"
     normalize: NormalizeArgs | Literal["auto"] = "auto"
@@ -92,9 +93,7 @@ class DINOv3EoMTSemanticSegmentationTrainTransformArgs(
     def resolve_auto(self, model_init_args: dict[str, Any]) -> None:
         super().resolve_auto(model_init_args=model_init_args)
         if self.image_size == "auto":
-            image_size = model_init_args.get("image_size", (518, 518))
-            assert isinstance(image_size, tuple)
-            self.image_size = image_size
+            self.image_size = tuple(model_init_args.get("image_size", (518, 518)))
 
         height, width = self.image_size
         for field_name in self.__class__.model_fields:
@@ -122,7 +121,7 @@ class DINOv3EoMTSemanticSegmentationValTransformArgs(SemanticSegmentationTransfo
     Defines default transform arguments for semantic segmentation validation with DINOv3.
     """
 
-    image_size: tuple[int, int] | Literal["auto"] = "auto"
+    image_size: ImageSizeTuple | Literal["auto"] = "auto"
     channel_drop: ChannelDropArgs | None = None
     num_channels: int | Literal["auto"] = "auto"
     normalize: NormalizeArgs | Literal["auto"] = "auto"
@@ -137,9 +136,7 @@ class DINOv3EoMTSemanticSegmentationValTransformArgs(SemanticSegmentationTransfo
     def resolve_auto(self, model_init_args: dict[str, Any]) -> None:
         super().resolve_auto(model_init_args=model_init_args)
         if self.image_size == "auto":
-            image_size = model_init_args.get("image_size", (518, 518))
-            assert isinstance(image_size, tuple)
-            self.image_size = image_size
+            self.image_size = tuple(model_init_args.get("image_size", (518, 518)))
 
         height, width = self.image_size
         for field_name in self.__class__.model_fields: