+++ CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add support for loading DICOM images as input data for training and inference.
-
 ### Changed
 
 ### Deprecated
@@ -19,7 +17,56 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix bug in `model.predict()` for object detection models.
+
 ### Security
+
+## [0.12.4] - 2025-11-26
+
+### Fixed
+
+- Fix bug in `model.predict()` for object detection models.
+
+## [0.12.3] - 2025-11-26
+
+### Added
+
+- Add support for specifying data configs in YAML format.
+
+### Changed
+
+- Improve the layout of the object detection training logs.
+
+### Deprecated
+
+- Deprecate `reuse_class_head` argument in `train` command. The model will now
+  automatically reuse the classification head only when the number of classes in the
+  data config matches that in the checkpoint. Otherwise, the classification head will
+  be re-initialized.
+
+### Fixed
+
+- Fix `image_size` not tuple when training from pretrained model.
+- Fix a bug when fine-tuning a model with `resume_interrupted=True`.
+- Fix `num_classes` not updated when loading an object detection checkpoint with
+  different number of classes.
+
+## [0.12.2] - 2025-11-14
+
+### Fixed
+
+- Fix `image_size` not tuple when training from pretrained model.
+
+## [0.12.1] - 2025-11-13
+
+### Added
+
+- Add support for DINOv3 [instance segmentation](https://docs.lightly.ai/train/stable/instance_segmentation.html)
+  inference and fine-tuning.
+- Add support for loading [DICOM images](https://docs.lightly.ai/train/stable/data/dicom.html)
+  as input data for training and inference.
+- Add event tracking, disable with `LIGHTLY_TRAIN_EVENTS_DISABLED=1`
+- Add support for fine-tuning object detection models with custom image resolutions.
 
 ## [0.12.0] - 2025-11-06